backend/requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
black==24.10.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.109.2
firebase-admin==6.8.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-secret-manager==2.25.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.26.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
pytest==7.4.4
pytest-asyncio==0.23.8
pytest-cov==4.1.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
ruff==0.1.15
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.36.3
tenacity==8.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
